Include spaces when checking guesses against answers
Before this commit brittbot would rip out spaces when cleaning guesses
and answers in trivia. This had an unintended effect of making all
guesses and answers a single word which broke the ignored words
functionality in trivia.

This commit also removes the "smartness" of adding in plural/non words
automatically.